Move the `verify_password_if_needed` helper up one level
